Cleaned up destruction process

+++ chunkdll.h
@@ -31,5 +31,5 @@ void __chunkdll_insert (struct chunknode* des, struct chunknode* c);
 struct chunknode* __chunkdll_pop (struct chunkdll* dll);//returns head. DOESN'T delete
 struct chunknode* __chunkdll_remove (struct chunknode* c);
 void __chunkdll_dump (struct chunkdll* dll, FILE* stream);
-unsigned int __chunkdll_destroy (struct chunkdll* dll);
+unsigned long __chunkdll_destroy (struct chunkdll* dll);
 #endif /* __CHUNKDLL_H__ */
